+++ tests/api/archive_multihost_test.rs
@@ -12,7 +12,7 @@ use actix_http::header::IF_NONE_MATCH;
 use actix_http::StatusCode;
 use actix_web::test;
 use std::path::PathBuf;
-use stelae::server::headers::HTTP_E_TAG;
+use taf_server::server::headers::HTTP_E_TAG;
 
 #[actix_web::test]
 async fn test_resolve_both_guarded_stele_law_html_request_with_full_path_expect_success() {