+++ tests/basic/archive_test.rs
@@ -1,4 +1,4 @@
-use stelae::utils::archive::find_archive_path;
+use taf_server::utils::archive::find_archive_path;
 
 use crate::common::{self, BASIC_MODULE_NAME};